fix: local upload commands should use file core api
The FolderUploadCommand.php as been renamed to UploadFolderCommand.php

- emsch:local:upload-assets
- emsch:local:upload-folder
